+++ controllers/expose/ingressClassCreator.go
@@ -1,17 +1,11 @@
-package controllers
+package expose
 
 import (
 	"context"
 	"fmt"
-	"k8s.io/client-go/tools/record"
-
-	appsv1 "k8s.io/api/apps/v1"
+	"github.com/cloudogu/k8s-service-discovery/controllers/util"
 	corev1 "k8s.io/api/core/v1"
-
-	"k8s.io/apimachinery/pkg/types"
-
 	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	networking "k8s.io/api/networking/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -22,27 +16,28 @@ const (
 	ingressClassCreationEventReason = "IngressClassCreation"
 )
 
-var k8sCesLabels = map[string]string{"app": "ces", "app.kubernetes.io/name": "k8s-service-discovery"}
-
-type eventRecorder interface {
-	record.EventRecorder
-}
-
 // ingressClassCreator is responsible to create a cluster wide ingress class in the cluster.
 type ingressClassCreator struct {
-	client        client.Client
-	className     string
-	namespace     string
-	eventRecorder eventRecorder
+	className             string
+	namespace             string
+	eventRecorder         eventRecorder
+	ingressController     ingressController
+	ingressClassInterface ingressClassInterface
+	deploymentInterface   deploymentInterface
 }
 
 // NewIngressClassCreator creates a new ingress class creator.
-func NewIngressClassCreator(client client.Client, className string, namespace string, recorder eventRecorder) *ingressClassCreator {
+func NewIngressClassCreator(clientset clientSetInterface, className string, namespace string, recorder eventRecorder, controller ingressController) *ingressClassCreator {
+	icInterface := clientset.NetworkingV1().IngressClasses()
+	deployInterface := clientset.AppsV1().Deployments(namespace)
+
 	return &ingressClassCreator{
-		client:        client,
-		className:     className,
-		namespace:     namespace,
-		eventRecorder: recorder,
+		className:             className,
+		namespace:             namespace,
+		eventRecorder:         recorder,
+		ingressController:     controller,
+		ingressClassInterface: icInterface,
+		deploymentInterface:   deployInterface,
 	}
 }
 
@@ -54,8 +49,7 @@ func (icc ingressClassCreator) CreateIngressClass(ctx context.Context) error {
 		return fmt.Errorf("failed to check if ingress class [%s] exists: %w", icc.className, err)
 	}
 
-	deployment := &appsv1.Deployment{}
-	err = icc.client.Get(ctx, types.NamespacedName{Name: "k8s-service-discovery-controller-manager", Namespace: icc.namespace}, deployment)
+	deployment, err := icc.deploymentInterface.Get(ctx, "k8s-service-discovery-controller-manager", metav1.GetOptions{})
 	if err != nil {
 		return fmt.Errorf("create ingress class: failed to get deployment [k8s-service-discovery-controller-manager]: %w", err)
 	}
@@ -70,14 +64,14 @@ func (icc ingressClassCreator) CreateIngressClass(ctx context.Context) error {
 		TypeMeta: metav1.TypeMeta{},
 		ObjectMeta: metav1.ObjectMeta{
 			Name:   icc.className,
-			Labels: k8sCesLabels,
+			Labels: util.K8sCesServiceDiscoveryLabels,
 		},
 		Spec: networking.IngressClassSpec{
-			Controller: "k8s.io/nginx-ingress",
+			Controller: icc.ingressController.GetControllerSpec(),
 		},
 	}
 
-	err = icc.client.Create(ctx, ingressClassResource)
+	_, err = icc.ingressClassInterface.Create(ctx, ingressClassResource, metav1.CreateOptions{})
 	if err != nil {
 		return fmt.Errorf("cannot create ingress class [%s] with clientset: %w", icc.className, err)
 	}
@@ -88,12 +82,7 @@ func (icc ingressClassCreator) CreateIngressClass(ctx context.Context) error {
 
 // isIngressClassAvailable check whether an ingress class with the given name exists in the current namespace.
 func (icc ingressClassCreator) isIngressClassAvailable(ctx context.Context) (bool, error) {
-	ingressClassKey := types.NamespacedName{
-		Namespace: "",
-		Name:      icc.className,
-	}
-	ingressClass := &networking.IngressClass{}
-	err := icc.client.Get(ctx, ingressClassKey, ingressClass)
+	_, err := icc.ingressClassInterface.Get(ctx, icc.className, metav1.GetOptions{})
 	if err != nil && apierrors.IsNotFound(err) {
 		return false, nil
 	} else if err != nil {