+++ controllers/expose/ingressController/interfaces.go
@@ -1,0 +1,7 @@
+package ingressController
+
+type IngressController interface {
+	GetControllerSpec() string
+	GetRewriteAnnotationKey() string
+	GetAdditionalConfigurationKey() string
+}