+++ controllers/util/interfaces.go
@@ -1,15 +1,10 @@
-package controllers
+package util
 
 import (
 	"context"
 	libconfig "github.com/cloudogu/k8s-registry-lib/config"
 	"github.com/cloudogu/k8s-registry-lib/repository"
-	"k8s.io/client-go/tools/record"
 )
-
-type eventRecorder interface {
-	record.EventRecorder
-}
 
 type GlobalConfigRepository interface {
 	Get(context.Context) (libconfig.GlobalConfig, error)